v0.94.0-alpha: Local (UNIX) socket support & queue view scroll fix
* Implement StreamWrapper

* Add local socket UI

* Resolve relative/tilde paths for local socket

* Improve connection error popups & clear state upon reconnecting

TODO: move folder view state to controller like other views.

* Fix playlist queue-all button queueing twice

* Fix queue view resetting scroll position on swaps

Scroll still resets on track removal. This seems hard to fix.

The biggest change of this commit is actually the whole queue update
logic, which is now local-first, server-later. We stay in sync with the
server by carefully monitoring queue version changes and tracking the
expected server-side version as changes are applied locally.

This new logic is much more efficient both for the server and the
client. It also allows us to at least get rid of a near-full refresh of
the queue upon song swaps and removals. Also thanks to song swaps being
performed locally first (and not having to be locally realised later via
a full queue refresh) we have gotten rid of the jump-to-beginning bug
at least for swaps.

For some reason removals will still reset the queue location. Not even a
manual scroll position restoration in g_idle_add_once() helped.

* Work around queue scroll position bug

* Refactor folder view state into library controller

Finishing what the background client refactor set out to do.

* Bump version and update metadata

* Fix Playlist Editor scroll issue too

Same issue as Queue View

+++ src/player/controller.rs
@@ -220,7 +220,7 @@ mod imp {
         pub fft_data: Arc<Mutex<(Vec<f32>, Vec<f32>)>>, // Binned magnitudes, in stereo
         pub use_visualizer: Cell<bool>,
         pub fft_backend_idx: Cell<i32>,
-        pub outputs: gio::ListStore
+        pub outputs: gio::ListStore,
     }
 
     #[glib::object_subclass]
@@ -275,7 +275,7 @@ mod imp {
                 ))),
                 use_visualizer: Cell::new(false),
                 fft_backend_idx: Cell::new(0),
-                outputs: gio::ListStore::new::<BoxedAnyObject>()
+                outputs: gio::ListStore::new::<BoxedAnyObject>(),
             }
         }
     }
@@ -638,18 +638,26 @@ impl Player {
                 #[weak(rename_to = this)]
                 self,
                 move |state, _| {
-                    if state.get_connection_state() == ConnectionState::Connected {
-                        // Newly-connected? Get initial status
-                        // Remember to get queue before status so status parsing has something to read off.
-                        if let Some(songs) = this.client().get_current_queue() {
-                            this.update_queue(&songs, true);
+                    match state.get_connection_state() {
+                        ConnectionState::Connected => {
+                            // Newly-connected? Get initial status
+                            // Remember to get queue before status so status parsing has something to read off.
+                            if let Some(songs) = this.client().get_current_queue() {
+                                this.update_queue(&songs, true);
+                            }
+                            if let Some(status) = this.client().get_status() {
+                                this.update_status(&status);
+                            }
+                            if let Some(outputs) = this.client().get_outputs() {
+                                this.update_outputs(outputs);
+                            }
                         }
-                        if let Some(status) = this.client().get_status() {
-                            this.update_status(&status);
+                        ConnectionState::Connecting => {
+                            this.imp().queue.remove_all();
+                            this.imp().outputs.remove_all();
+                            this.update_status(&mpd::Status::default());
                         }
-                        if let Some(outputs) = this.client().get_outputs() {
-                            this.update_outputs(outputs);
-                        }
+                        _ => {}
                     }
                 }
             ),
@@ -667,16 +675,7 @@ impl Player {
                 self,
                 move |_: ClientState, subsys: glib::BoxedAnyObject| {
                     match subsys.borrow::<Subsystem>().deref() {
-                        Subsystem::Player | Subsystem::Options => {
-                            if let Some(status) = this.client().get_status() {
-                                this.update_status(&status);
-                            }
-                        }
-                        Subsystem::Queue => {
-                            if let Some(songs) = this.client().get_queue_changes() {
-                                this.update_queue(&songs, false);
-                            }
-                            // Need to also update queue length
+                        Subsystem::Player | Subsystem::Queue | Subsystem::Options => {
                             if let Some(status) = this.client().get_status() {
                                 this.update_status(&status);
                             }
@@ -685,9 +684,21 @@ impl Player {
                             if let Some(outs) = this.client().get_outputs() {
                                 this.update_outputs(outs);
                             }
-                        }
+                        } 
                         _ => {}
                     }
+                }
+            ),
+        );
+
+        client_state.connect_closure(
+            "queue-changed",
+            false,
+            closure_local!(
+                #[weak(rename_to = this)]
+                self,
+                move |_: ClientState, songs: BoxedAnyObject| {
+                    this.update_queue(songs.borrow::<Vec<Song>>().as_ref(), false);
                 }
             ),
         );
@@ -1041,9 +1052,9 @@ impl Player {
     /// before update_status() (by MPD's idle change notifier) and as such has no way to know the
     /// new queue length. The update_status() function will instead truncate the queue to the new
     /// length for us once called.
-    ///
     /// If an MPRIS server is running, it will also emit property change signals.
-    pub fn update_queue(&self, songs: &[Song], replace: bool) {
+    pub fn update_queue(&self, songs: &[Song], replace: bool) { 
+        println!("RUNNING UPDATE_QUEUE");
         let queue = &self.imp().queue;
         if replace {
             if songs.len() == 0 {
@@ -1104,7 +1115,6 @@ impl Player {
             // Might queue downloads, depending on user settings
             cache.ensure_cached_album_arts(&infos);
         }
-        // Downstream widgets should now receive an item-changed signal.
     }
 
     fn client(&self) -> &Rc<MpdWrapper> {
@@ -1336,20 +1346,44 @@ impl Player {
         self.client().play_at(song.get_queue_id(), true);
     }
 
-    pub fn remove_song_id(&self, id: u32) {
-        self.client().delete_at(id, true);
+    fn pos_of_id(&self, id: u32) -> Option<u32> {
+        for (i, song_obj) in self.imp().queue.iter::<Song>().enumerate() {
+            let song: Song = song_obj.unwrap().downcast::<Song>().unwrap();
+            if song.get_queue_id() == id {
+                return Some(i as u32);
+            }
+        }
+        None
     }
 
-    pub fn swap_dir(&self, pos: u32, direction: SwapDirection) {
-        match direction {
-            SwapDirection::Up => {
-                if pos > 0 {
-                    self.client().swap(pos, pos - 1, false);
+    pub fn remove_song_id(&self, id: u32) {
+        if let Some(pos) = self.pos_of_id(id) {
+            self.client().register_local_queue_changes(1);
+            self.imp().queue.remove(pos);
+            self.client().delete_at(id, true);
+        }
+    }
+
+    pub fn swap_dir(&self, id: u32, direction: SwapDirection) {
+        // Find position of given queue_id
+        if let Some(pos) = self.pos_of_id(id) {
+            self.client().register_local_queue_changes(1);
+            match direction {
+                SwapDirection::Up => {
+                    if pos > 0 {
+                        let target = self.imp().queue.item(pos).unwrap();
+                        let upper = self.imp().queue.item(pos - 1).unwrap();
+                        self.imp().queue.splice(pos - 1, 2, &[target, upper]);
+                        self.client().swap(pos, pos - 1, false);
+                    }
                 }
-            }
-            SwapDirection::Down => {
-                if pos < self.imp().queue.n_items() - 1 {
-                    self.client().swap(pos, pos + 1, false);
+                SwapDirection::Down => {
+                    if pos < self.imp().queue.n_items() - 1 {
+                        let target = self.imp().queue.item(pos).unwrap();
+                        let lower = self.imp().queue.item(pos + 1).unwrap();
+                        self.imp().queue.splice(pos, 2, &[lower, target]);
+                        self.client().swap(pos, pos + 1, false);
+                    }
                 }
             }
         }
